requirements.txt
Django==5.1.3
tempmail-python==2.3.3